+++ python/ittapi.native/extensions/string.cpp
