+++ python/ittapi.native/extensions/string.hpp
