+++ python/ittapi.native/string_handle.hpp
