+++ python/ittapi/collection_control.py
@@ -1,7 +1,7 @@
 """
 collection_control.py - Python module wrapper for ITT Collection Control API
 """
-from pyitt.native import detach as _detach, pause as _pause, resume as _resume
+from ittapi.native import detach as _detach, pause as _pause, resume as _resume
 
 from .region import _Region
