+++ python/ittapi/thread_naming.py
@@ -1,7 +1,7 @@
 """
 thread_naming.py - Python module wrapper for ITT Thread Naming API
 """
-from pyitt.native import thread_set_name as _thread_set_name
+from ittapi.native import thread_set_name as _thread_set_name
 
 
 def thread_set_name(name: str):
