+++ python/pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "intel-pyitt"
+name = "intel-ittapi"
 version = "1.1.0"
 authors = [
   { name="Egor Suldin", email="egor.suldin@intel.com" },
