+++ python/samples/event_sample.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-import pyitt
+import ittapi
 
 # pylint: disable=C0411
 from argparse import ArgumentParser
@@ -8,31 +8,31 @@ from workload import workload
 
 
 def run_sample():
-    # pyitt.event can be used as decorator
-    @pyitt.event
+    # ittapi.event can be used as decorator
+    @ittapi.event
     def my_function_1():
         workload()
 
     # the list of arguments can be empty
-    @pyitt.event()
+    @ittapi.event()
     def my_function_2():
         workload()
 
     # or you can specify the name of the event and other parameters
-    @pyitt.event('my function 3')
+    @ittapi.event('my function 3')
     def my_function_3():
         workload()
 
-    @pyitt.event
-    @pyitt.event('my function 4')
+    @ittapi.event
+    @ittapi.event('my function 4')
     def my_function_4():
         workload()
 
-    # also, pyitt.event can be used as a context manager
-    with pyitt.event():
+    # also, ittapi.event can be used as a context manager
+    with ittapi.event():
         workload()
     # in this form you also can specify the name, the domain and other parameters in the same way
-    with pyitt.event('my event'):
+    with ittapi.event('my event'):
         workload()
 
     my_function_1()
@@ -41,10 +41,10 @@ def run_sample():
     my_function_4()
 
     # example for overlapped events
-    overlapped_event_1 = pyitt.event('overlapped event 1')
+    overlapped_event_1 = ittapi.event('overlapped event 1')
     overlapped_event_1.begin()
     workload()
-    overlapped_event_2 = pyitt.event('overlapped event 2')
+    overlapped_event_2 = ittapi.event('overlapped event 2')
     overlapped_event_2.begin()
     workload()
     overlapped_event_1.end()
@@ -56,7 +56,7 @@ def run_sample():
         def __call__(self, *args, **kwargs):  # pylint: disable=W0621
             workload()
 
-    callable_object = pyitt.event(CallableClass())
+    callable_object = ittapi.event(CallableClass())
     callable_object()
