+++ python/samples/task_sample.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-import pyitt
+import ittapi
 
 # pylint: disable=C0411
 from argparse import ArgumentParser
@@ -8,36 +8,36 @@ from workload import workload
 
 
 def run_sample():
-    # pyitt.task can be used as decorator
-    @pyitt.task
+    # ittapi.task can be used as decorator
+    @ittapi.task
     def my_function_1():
         workload()
 
     # the list of arguments can be empty
-    @pyitt.task()
+    @ittapi.task()
     def my_function_2():
         workload()
 
     # or you can specify the name of the task and other parameters
-    @pyitt.task('my function 3')
+    @ittapi.task('my function 3')
     def my_function_3():
         workload()
 
     # like domain
-    @pyitt.task(domain='my domain')
+    @ittapi.task(domain='my domain')
     def my_function_4():
         workload()
 
-    @pyitt.task
-    @pyitt.task('my function 5')
+    @ittapi.task
+    @ittapi.task('my function 5')
     def my_function_5():
         workload()
 
-    # also, pyitt.task can be used as a context manager
-    with pyitt.task():
+    # also, ittapi.task can be used as a context manager
+    with ittapi.task():
         workload()
     # in this form you also can specify the name, the domain and other parameters in the same way
-    with pyitt.task('my task', 'my domain'):
+    with ittapi.task('my task', 'my domain'):
         workload()
 
     my_function_1()
@@ -47,10 +47,10 @@ def run_sample():
     my_function_5()
 
     # example for overlapped tasks
-    overlapped_task_1 = pyitt.task('overlapped task 1', overlapped=True)
+    overlapped_task_1 = ittapi.task('overlapped task 1', overlapped=True)
     overlapped_task_1.begin()
     workload()
-    overlapped_task_2 = pyitt.task('overlapped task 2', overlapped=True)
+    overlapped_task_2 = ittapi.task('overlapped task 2', overlapped=True)
     overlapped_task_2.begin()
     workload()
     overlapped_task_1.end()
@@ -62,7 +62,7 @@ def run_sample():
         def __call__(self, *args, **kwargs):  # pylint: disable=W0621
             workload()
 
-    callable_object = pyitt.task(CallableClass())
+    callable_object = ittapi.task(CallableClass())
     callable_object()
