+++ python/utest/test_collection_control.py
@@ -1,23 +1,23 @@
 from unittest import main as unittest_main, TestCase
 
 from pyitt_native_mock import patch as pyitt_native_patch
-import pyitt
+import ittapi
 
 
 class DirectCollectionControlTests(TestCase):
     @pyitt_native_patch('detach')
     def test_detach_call(self, detach_mock):
-        pyitt.collection_control.detach()
+        ittapi.collection_control.detach()
         detach_mock.assert_called_once()
 
     @pyitt_native_patch('pause')
     def test_pause_call(self, pause_mock):
-        pyitt.collection_control.pause()
+        ittapi.collection_control.pause()
         pause_mock.assert_called_once()
 
     @pyitt_native_patch('resume')
     def test_resume_call(self, resume_mock):
-        pyitt.collection_control.resume()
+        ittapi.collection_control.resume()
         resume_mock.assert_called_once()
 
 
@@ -25,7 +25,7 @@ class ActiveRegionTests(TestCase):
     @pyitt_native_patch('pause')
     @pyitt_native_patch('resume')
     def test_active_region_as_decorator(self, pause_mock, resume_mock):
-        @pyitt.active_region
+        @ittapi.active_region
         def my_function():
             return 42
 
@@ -36,7 +36,7 @@ class ActiveRegionTests(TestCase):
     @pyitt_native_patch('pause')
     @pyitt_native_patch('resume')
     def test_active_region_as_context_manager(self, pause_mock, resume_mock):
-        with pyitt.active_region():
+        with ittapi.active_region():
             pass
 
         resume_mock.assert_called_once()
@@ -45,7 +45,7 @@ class ActiveRegionTests(TestCase):
     @pyitt_native_patch('pause')
     @pyitt_native_patch('resume')
     def test_active_region_with_manual_activation(self, pause_mock, resume_mock):
-        region = pyitt.active_region()
+        region = ittapi.active_region()
 
         region.activator.deactivate()
         with region:
@@ -65,7 +65,7 @@ class ActiveRegionTests(TestCase):
     @pyitt_native_patch('resume')
     def test_active_region_with_custom_activator(self, pause_mock, resume_mock):
         for i in range(4):
-            with pyitt.active_region(activator=lambda: i % 2):  # pylint: disable=W0640
+            with ittapi.active_region(activator=lambda: i % 2):  # pylint: disable=W0640
                 pass
 
         self.assertEqual(resume_mock.call_count, 2)
@@ -74,7 +74,7 @@ class ActiveRegionTests(TestCase):
     @pyitt_native_patch('pause')
     @pyitt_native_patch('resume')
     def test_active_region_as_decorator_without_activator(self, pause_mock, resume_mock):
-        @pyitt.active_region(activator=None)
+        @ittapi.active_region(activator=None)
         def my_function():
             return 42
 
@@ -87,7 +87,7 @@ class PausedRegionTests(TestCase):
     @pyitt_native_patch('pause')
     @pyitt_native_patch('resume')
     def test_paused_region_as_decorator(self, pause_mock, resume_mock):
-        @pyitt.paused_region
+        @ittapi.paused_region
         def my_function():
             return 42
 
@@ -98,7 +98,7 @@ class PausedRegionTests(TestCase):
     @pyitt_native_patch('pause')
     @pyitt_native_patch('resume')
     def test_paused_region_as_context_manager(self, pause_mock, resume_mock):
-        with pyitt.paused_region():
+        with ittapi.paused_region():
             pass
 
         resume_mock.assert_called_once()
@@ -107,7 +107,7 @@ class PausedRegionTests(TestCase):
     @pyitt_native_patch('pause')
     @pyitt_native_patch('resume')
     def test_paused_region_with_manual_activation(self, pause_mock, resume_mock):
-        region = pyitt.paused_region()
+        region = ittapi.paused_region()
 
         region.activator.deactivate()
         with region:
@@ -127,7 +127,7 @@ class PausedRegionTests(TestCase):
     @pyitt_native_patch('resume')
     def test_paused_region_with_custom_activator(self, pause_mock, resume_mock):
         for i in range(4):
-            with pyitt.paused_region(activator=lambda: i % 2):  # pylint: disable=W0640
+            with ittapi.paused_region(activator=lambda: i % 2):  # pylint: disable=W0640
                 pass
 
         self.assertEqual(resume_mock.call_count, 2)
@@ -136,7 +136,7 @@ class PausedRegionTests(TestCase):
     @pyitt_native_patch('pause')
     @pyitt_native_patch('resume')
     def test_paused_region_as_decorator_without_activator(self, pause_mock, resume_mock):
-        @pyitt.paused_region(activator=None)
+        @ittapi.paused_region(activator=None)
         def my_function():
             return 42
